Development in progress
Added the casualty to be black or white. Now the FEN string is still generated in a .txt file but it is quickly deleted and saved in the local storage so if the user closes the browser he can recover his match. Resolved the problem for the files permanently. Started to design the model and the controller for the login and the user