Development in progress
Added the puzzle section. It makes a get request to the lichess.org API to get the information needed to perform the task. The python script changed so that now it will return only the move and not the fen string. By doing so it is now possible to use the chess.js function chess.pgn() that return the pgn file format of the match