Development in progress
Found a major problem with the promotion dialog on mobile device. Added arrows and circle when the user right click the chessboard. Added the puzzle model. The idea is to store all the puzzle in the database bbut only the id and other small stuff so that the lichess.api can be used to retreive all the information about the puzzle.