Development in progress
Added google autentication. Next thing to do will be to add the user who logged in with google in the database. Changed also the way the dispatcher works. It now calls the right controller based on the action. The controller has function such has index() that will require the right view.